Reorganise package structure
