+++ tv/src/main/java/au/com/shiftyjelly/pocketcasts/TvActivity.kt
@@ -3,6 +3,7 @@ package au.com.shiftyjelly.pocketcasts
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import au.com.shiftyjelly.pocketcasts.home.TvScaffold
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
